+++ lib/maze/capabilities.rb
@@ -9,14 +9,15 @@ module Maze
       # @param capabilities_option [String] extra capabilities provided on the command line
       def for_browser_stack_device(device_type, local_id, appium_version, capabilities_option)
         capabilities = {
-          'browserstack.console' => 'errors',
-          'browserstack.localIdentifier' => local_id,
-          'browserstack.local' => 'true',
+          'bstack:options' => {
+            'local' => 'true',
+            'localIdentifier' => local_id
+          },
           'noReset' => 'true'
         }
-        capabilities.merge! BrowserStackDevices::DEVICE_HASH[device_type]
-        capabilities.merge! JSON.parse(capabilities_option)
-        capabilities['browserstack.appium_version'] = appium_version unless appium_version.nil?
+        capabilities.deep_merge! BrowserStackDevices::DEVICE_HASH[device_type]
+        capabilities.deep_merge! JSON.parse(capabilities_option)
+        capabilities['bstack:options']['appiumVersion'] = appium_version unless appium_version.nil?
         capabilities
       end
 
@@ -24,14 +25,18 @@ module Maze
       # @param local_id [String] unique key for the tunnel instance
       # @param capabilities_option [String] extra capabilities provided on the command line
       def for_browser_stack_browser(browser_type, local_id, capabilities_option)
-        capabilities = Selenium::WebDriver::Remote::Capabilities.new
-        capabilities['browserstack.local'] = 'true'
-        capabilities['browserstack.localIdentifier'] = local_id
-        capabilities['browserstack.console'] = 'errors'
+        capabilities = {
+          'bstack:options' => {
+            'local' => 'true',
+            'localIdentifier' => local_id,
+            "os" => "Windows",
+            "osVersion" => "8.1"
+          }
+        }
         browsers = YAML.safe_load(File.read("#{__dir__}/browsers_bs.yml"))
-        capabilities.merge! browsers[browser_type]
-        capabilities.merge! JSON.parse(capabilities_option)
-        capabilities
+        capabilities.deep_merge! browsers[browser_type]
+        capabilities.deep_merge! JSON.parse(capabilities_option)
+        Selenium::WebDriver::Remote::Capabilities.new capabilities
       end
 
       # @param browser_type [String] A key from @see browsers_cbt.yml